Add support for capabilities substitutions
Similarly to dependency attributes substitutions, this commit adds
support for the other aspect of variant-aware selection which is
capabilities support.

It's now possible to substitute a dependency with or without explicit
capabilities with another dependency with or without capabilities.

For composite builds, the requested capabilities are _always_ used,
independently of what the user declares (in case of a mapping definition)
or not (implicit mapping).